pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chameleon-usage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ibis-framework[mysql]>=11.0.0",
    "matplotlib>=3.10.8",
    "pandera[polars]>=0.28.1",
    "polars>=1.37.1",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
